% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{box.dat}
\alias{box.dat}
\title{Information of boxes and layers}
\usage{
box.dat(bgm.file, cum.depths)
}
\arguments{
\item{bgm.file}{BGM file for atlantis}

\item{cum.depths}{cummulative depth}
}
\value{
A data.frame with the layers information
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}
