% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{bypol.eddy}
\alias{bypol.eddy}
\title{Eddy by polygon}
\usage{
bypol.eddy(eddy.file, coor.dat, polygons.dat, cyclonic = TRUE,
  relocate = NULL, centroids = NULL)
}
\arguments{
\item{eddy.file}{File that contain the trajectory of the eddies}

\item{coor.dat}{coordinates of the realm (lat and lon of domain)}

\item{polygons.dat}{Polygons of the model}

\item{cyclonic}{Logic (TRUE or FALSE) to stablish if the eddy is cyclonic or not}

\item{relocate}{Areas with no eddyes,  to avoid land and border polygons}

\item{centroids}{Centroids of polygon to relocate eddies located on land}
}
\value{
Number of eddies by polygon in Atlantis
}
\description{
Function to identified the number of eddies by polygons
}
\author{
Demiurgo
}
