% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{calc.edd}
\alias{calc.edd}
\title{Calculate the number of eddies by box and timestep}
\usage{
calc.edd(eddy.file, coor.dat, polygons.dat, cyclonic = FALSE, t.time,
  by.area = FALSE)
}
\arguments{
\item{eddy.file}{R.data file with the lat lon information}

\item{coor.dat}{Coordinates data file}

\item{polygons.dat}{Polygons information}

\item{cyclonic}{Type of eddie}
}
\value{
A matrix array(box, time) with the number of eddies by box
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}
