% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{deg2rad}
\alias{deg2rad}
\title{Degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Coordinates in degrees}
}
\value{
Object with the Lat and Lon information in Radians
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}
