% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{dilat}
\alias{dilat}
\title{Dilat borders}
\usage{
dilat(matrix)
}
\arguments{
\item{matrix}{Matrix witht he location of the border of the eddy}
}
\value{
The location of the border of the eddy after dilatation of the initial border
}
\description{
{} dilat the matrix in three cell in avery direction, its using 3
    cells because the eddies are moving 2km per day. in 5 days is almost 10 km,
    and that is roughly 3 cells. The maximum space that they can move
}
\author{
Demiurgo
}
