% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{edd.pos}
\alias{edd.pos}
\title{Patern recognition}
\usage{
edd.pos(eddies.pos, distx = NA, disty = NA, min.area = 100)
}
\arguments{
\item{eddies.pos}{Position of the eddy}

\item{distx}{Distance in x (m)}

\item{disty}{Distance in y (m)}

\item{min.area}{Minimum area for the eddie,  depends on the numbers of cells}
}
\value{
The position of the eddy,  the total area and the number of cells
}
\description{
This function discriminate eddies in a time-varying spatial field
    using an image processing algorithm based on the recognition of bi-tone (0 -
    1) patterns over grid cells.
}
\author{
Demiurgo
}
