% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{eddies.year}
\alias{eddies.year}
\title{Eddies by Year}
\usage{
eddies.year(files, folder)
}
\arguments{
\item{files}{Vector witht the names of the netcdf files}

\item{folder}{Folder that contain the folders}
}
\value{
The location of the border of the eddy (border cells) and information of
    the eddy (Cyclonic and Anticyclonic)
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}
