% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{eddy.bord}
\alias{eddy.bord}
\title{Eddy border}
\usage{
eddy.bord(eddy)
}
\arguments{
\item{eddy}{eddy information with the distance in x and y and the area of the eddy}
}
\value{
The borders of the eddy
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}
