% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{eddy.track}
\alias{eddy.track}
\title{Eddy tracker}
\usage{
eddy.track(file, cyclonic = TRUE, min.temp = 3)
}
\arguments{
\item{file}{Object witht the location of the eddy}

\item{cyclonic}{Logic (TRUE or FALSE) to stablish if the eddy is cyclonic or not}

\item{min.temp}{Minimum time duration of the eddy in the area}
}
\value{
The Eddies with a duration longer that  *min.temp*
}
\description{
track the eddies in base a standar distance of 3 cell (10 - 15 km)
}
\author{
Demiurgo
}
