% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{finder}
\alias{finder}
\title{finder}
\usage{
finder(org.edd, step2, tid)
}
\arguments{
\item{org.edd}{Position of the original eddy}

\item{step2}{Position of all the eddies in the next time step}

\item{tid}{ID that identified the eddy}
}
\value{
The next position of the eddy
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}
