% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{gcd.hf}
\alias{gcd.hf}
\title{Haversine distance formula}
\usage{
gcd.hf(long1 = NULL, lat1 = NULL, long2 = NULL, lat2 = NULL,
  mat = FALSE, pos1 = NULL, pos2 = NULL)
}
\arguments{
\item{long1}{Longitude of the first point in radians}

\item{lat1}{Latitude of the second point in radians}

\item{long2}{Longitude of the first point in radians}

\item{lat2}{Latitude of the second point in radians}

\item{mat}{TRUE or FALSE,  identified if the information is provided in two diferent matrix, one for the initial position and a second matrix for the final position}

\item{pos1}{If mat = TRUE this matrix represent the first position witht the first colum for the Longitude and the second for the Latitude}

\item{pos2}{If mat = TRUE this matrix represent the second position witht the first colum for the Longitude and the second for the Latitude}
}
\value{
the distance in m for eatch points
}
\description{
Calculates the geodesic distance between two points specified by radian latitude/longitude using the Haversine formula (hf)
}
\author{
Demiurgo
}
