% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{make.eddy.nc}
\alias{make.eddy.nc}
\title{Create the Netcdf file for eddies}
\usage{
make.eddy.nc(bgm.file, cum.depths, total.edd, nc.file)
}
\arguments{
\item{bgm.file}{## BGm file from atlatnis}

\item{cum.depths}{Cummulative depths of the layers}

\item{total.edd}{Matrix with the number of eddies by box (rows) by time step (cols)}

\item{nc.file}{Name of the ncfile outpur}
}
\value{
nc.file saved
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}
