% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{move.eddy}
\alias{move.eddy}
\title{Relocation of eddies from inside of a polygon to another closest}
\usage{
move.eddy(eddies, centroids, polygon)
}
\arguments{
\item{eddies}{Location of eddies}

\item{centroids}{Centroids of the polygons}

\item{polygon}{IDPolygons to remove}
}
\value{
the new position of the eddies
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Demiurgo
}
