% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{okubo}
\alias{okubo}
\title{Okuvo - Weiss Formula}
\usage{
okubo(lat, lon, u, v)
}
\arguments{
\item{lat}{Latitudes}

\item{lon}{Longitudes}

\item{u}{Velocity field in x (cm s-1)}

\item{v}{Velocity field in y (cm s-1)}
}
\value{
a list witht the values for the relative vorticity, the okuvo - weiss
    parameter value, the distance in x and the distance in y
}
\description{
Calculate the Okubo-Weiss parameter, return the normal component,
    zisalle and vorticity, calculated from the velocity fields (u, v), for each
    point in space. based on Correa-Ramirez et al. (2007)
}
\author{
Demiurgo
}
