% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{sigma2zeta}
\alias{sigma2zeta}
\title{Sigma to meters}
\usage{
sigma2zeta(h, hc, theta, b, N)
}
\arguments{
\item{h}{depth in m}

\item{hc}{20}

\item{theta}{5}

\item{b}{0.6}

\item{N}{Number of Layers}
}
\value{
return the depth (z) in meters
}
\description{
Calculation fo the depth in z - coordinates (meters) from a sigma
    layer coodinates. The stimation is do it for the center of the polygon
}
\author{
Demiurgo
}
