% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{uv2rho}
\alias{uv2rho}
\title{Transfer to Rho points}
\usage{
uv2rho(var, varname = NULL)
}
\arguments{
\item{var}{Vector with the values}

\item{varname}{Name of the variable, it can be U or V}
}
\value{
Returns the vector in in Rho points
}
\description{
This function transfert a field at u or v points to a field at rho points
}
\author{
Demiurgo
}
