% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddies.R
\name{w.area}
\alias{w.area}
\title{Weighted by area}
\usage{
w.area(n.eddies, area.rm = NULL)
}
\arguments{
\item{n.eddies}{Number of eddies by polygon}

\item{area.rm}{Area of the polygon (m2)}
}
\value{
A weighted number of eddies
}
\description{
This function weight the number of eddies by the area of the polygon

.. content for \details{} ..
}
\author{
Demiurgo
}
